No longer add version into generated verification metadata files
This means that the autogenerated verification metadata files will say that a certain key is valid for a group and artifact rather than only for a specific version of that group and artifact.

Fixes: #20192
